Overnight / TA Assisted work - Debugged and Converted Routes.
Debugged and converted user and thought routes to be async functions with try catch blocks instead.  Working with the TA from class, he was able to assist me with converting over a function as an example to assist with previous errors.  This fixed the issues with null objects being passed over and allowed for proper error catching.  Additionally, this fixed issues with proper response codes and objects showing in insomnia when making calls.